Added ViewComposer middleware. It will be used to inject some code after app language was set
