+++ src/Gzero/Core/Http/Middleware/MultiLanguage.php
@@ -8,31 +8,6 @@ use Gzero\DomainException;
 
 class MultiLanguage {
 
-    /** @var array */
-    protected static $callbacks = [];
-
-    /**
-     * It adds callback to callbacks array
-     *
-     * @param Closure $callback Callback
-     *
-     * @return void
-     */
-    public static function addCallback(Closure $callback)
-    {
-        static::$callbacks[] = $callback;
-    }
-
-    /**
-     * It clears callbacks array
-     *
-     * @return void
-     */
-    public static function clearCallbacks()
-    {
-        static::$callbacks = [];
-    }
-
     /**
      * Handle an incoming request.
      *
@@ -79,12 +54,6 @@ class MultiLanguage {
         }
 
         resolve(RoutesService::class)->registerAll();
-
-        // We use this callbacks to register some view shares after we change app locale
-        foreach (static::$callbacks as $callback) {
-            $callback();
-            static::clearCallbacks();
-        }
 
         return $next($request);
     }
