+++ src/Gzero/Core/ServiceProvider.php
@@ -1,6 +1,7 @@
 <?php namespace Gzero\Core;
 
 use Carbon\Carbon;
+use Gzero\Core\Http\Middleware\ViewComposer;
 use Gzero\Core\Policies\FilePolicy;
 use Gzero\Core\Http\Middleware\Init;
 use Gzero\Core\Http\Middleware\MultiLanguage;
@@ -216,6 +217,7 @@ class ServiceProvider extends AbstractServiceProvider {
         /** @var Router $router */
         $router = resolve(Router::class);
         $router->pushMiddlewareToGroup('web', CreateFreshApiToken::class);
+        $router->pushMiddlewareToGroup('web', ViewComposer::class);
         $router->pushMiddlewareToGroup('web', ViewShareUser::class);
     }
